+++ user-service/src/main/java/com/flight_booking/user_service/application/service/BlockService.java
@@ -3,7 +3,7 @@ package com.flight_booking.user_service.application.service;
 import com.flight_booking.user_service.domain.model.Role;
 import com.flight_booking.user_service.domain.model.User;
 import com.flight_booking.user_service.domain.repository.UserRepository;
-import com.flight_booking.user_service.infrastructure.security.authentication.CustomUserDetails;
+import com.flight_booking.user_service.infrastructure.security.CustomUserDetails;
 import com.flight_booking.user_service.presentation.global.exception.ErrorCode;
 import com.flight_booking.user_service.presentation.global.exception.UserException;
 import com.flight_booking.user_service.presentation.request.BlockRequest;
@@ -45,7 +45,7 @@ public class BlockService {
   // 존재하는 사용자 확인 및 삭제된 사용자 확인
   private User getUser(Long id) {
     User user = userRepository.findById(id)
-        .orElseThrow(() -> new UserException(ErrorCode.ACCESS_ONLY_SELF));
+        .orElseThrow(() -> new UserException(ErrorCode.USER_NOT_FOUND));
 
     if (user.getIsDeleted()) {
       throw new UserException(ErrorCode.USER_DELETED);
